add score change

+++ Src/Runtime/Module/Instancing/Data/PlayerInstancingData.cs
@@ -14,8 +14,8 @@ public class PlayerInstancingData
     public bool IsSettlement; //是否结算
     public bool IsDeductTicket; //是否扣除门票
     public HashSet<int> CompleteLevelList = new(); //已经完成的关卡
-    public int Score;                              //副本分数
-    public int LastLevelScore;                    //上一关卡的分数
+    public float Score;                              //副本分数
+    public float LastLevelScore;                     //上一关卡的分数
     public float TotemScore;                       //图腾分数
     public int LifeCount; //生命数量 
     public bool IsWaitTransfer; //是否等待传送
@@ -44,7 +44,7 @@ public class PlayerInstancingData
     public GameMessageCore.InstancingPlayerData GetNetData()
     {
         NetData.PlayerId = PlayerId;
-        NetData.Score = Score;
+        NetData.Score = (int)Score;
         NetData.TotemScore = TotemScore;
         NetData.LifeCount = LifeCount;
         return NetData;
